+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## UNRELEASED
 
+- 💥 BREAKING: Disable all `autofill` annotations by default.
+  This can be reverted by setting the `HERMAN_ENABLE_AUTOFILL`
+  environment variable to a truthy value
+  (e.g. `HERMAN_ENABLE_AUTOFILL=1 sassdoc`).
 - 💥 BREAKING: Drop support for Node < 12
 - 💥 BREAKING: `utilities.add()`
   no longer supports map-compilation functions and arguments,